+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Release to Github Packages
+- Explicitly flush datadog metrics after publishing them
+
+### Changed
+- Replace `statsd.batch` with direct calls, as it aggregates metrics interally by default now.
+  Also `#batch` method is deprecated and will be removed in version 6 of Datadog Statsd client.
 
 ## [0.3.1] - 2021-03-26
 ### Changed