+++ CHANGELOG.md
@@ -6,9 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.0.0] - 2021-09-08
 ### Added
 - Release to Github Packages
 - Explicitly flush datadog metrics after publishing them
+- Middleware for measuring and tracking request queue time
 
 ### Changed
 - Replace `statsd.batch` with direct calls, as it aggregates metrics interally by default now.